+++ examples/internal_link_sets.py
@@ -1,6 +1,6 @@
 # -*- coding: utf-8 -*-
 ###
-# (C) Copyright (2012-2017) Hewlett Packard Enterprise Development LP
+# (C) Copyright (2012-2019) Hewlett Packard Enterprise Development LP
 #
 # Permission is hereby granted, free of charge, to any person obtaining a copy
 # of this software and associated documentation files (the "Software"), to deal
@@ -28,29 +28,24 @@ from config_loader import try_load_from_file
 config = {
     "ip": "<oneview_ip>",
     "credentials": {
-        "userName": "<oneview_administrator_name>",
-        "password": "<oneview_administrator_password>",
+        "userName": "<username>",
+        "password": "<password>",
     }
 }
 
 # Try load config from a file (if there is a config file)
 config = try_load_from_file(config)
-
 oneview_client = OneViewClient(config)
+internal_link_sets = oneview_client.internal_link_sets
 
 # Get all, with defaults
 print("Get all internal-link-sets")
-internal_links = oneview_client.internal_link_sets.get_all()
+internal_links = internal_link_sets.get_all()
 pprint(internal_links)
 
 internal_link_set_uri = internal_links[0]['uri']
 internal_link_set_name = internal_links[0]['name']
 
-# Get by Uri
-print("\nGet a internal-link-set by uri")
-internal_links_by_uri = oneview_client.internal_link_sets.get(internal_link_set_uri)
-pprint(internal_links_by_uri)
-
 # Find by name
-internal_links_by_name = oneview_client.internal_link_sets.get_by('name', internal_link_set_name)
-print("\nFound the internal-link-sets by name: '{}':".format(internal_links_by_name[0]['name']))
+internal_links_by_name = internal_link_sets.get_by_name(internal_links[0]["name"])
+print("\nFound the internal-link-sets by name: '{}':".format(internal_links_by_name.data["name"]))