+++ tests/unit/resources/networking/test_sas_interconnect_types.py
@@ -1,6 +1,6 @@
 # -*- coding: utf-8 -*-
 ###
-# (C) Copyright (2012-2017) Hewlett Packard Enterprise Development LP
+# (C) Copyright (2012-2019) Hewlett Packard Enterprise Development LP
 #
 # Permission is hereby granted, free of chlarge, to any person obtaining a copy
 # of this software and associated documentation files (the "Software"), to deal
@@ -27,7 +27,7 @@ import mock
 
 from hpOneView.connection import connection
 from hpOneView.resources.networking.sas_interconnect_types import SasInterconnectTypes
-from hpOneView.resources.resource import ResourceClient
+from hpOneView.resources.resource import ResourceHelper
 
 
 class SasInterconnectTypesTest(unittest.TestCase):
@@ -36,28 +36,9 @@ class SasInterconnectTypesTest(unittest.TestCase):
         self.connection = connection(self.host)
         self._sas_interconnect_types = SasInterconnectTypes(self.connection)
 
-    @mock.patch.object(ResourceClient, 'get_all')
+    @mock.patch.object(ResourceHelper, 'get_all')
     def test_get_all_called_once(self, mock_get_all):
         filter = 'name=TestName'
         sort = 'name:ascending'
 
         self._sas_interconnect_types.get_all(2, 500, filter, sort)
-
-    @mock.patch.object(ResourceClient, 'get_by')
-    def test_get_by_called_once(self, mock_get_by):
-        self._sas_interconnect_types.get_by('name', 'SAS Interconnect Type 1')
-
-        mock_get_by.assert_called_once_with('name', 'SAS Interconnect Type 1')
-
-    @mock.patch.object(ResourceClient, 'get')
-    def test_get_called_once(self, mock_get):
-        self._sas_interconnect_types.get('3518be0e-17c1-4189-8f81-83f3724f6155')
-
-        mock_get.assert_called_once_with('3518be0e-17c1-4189-8f81-83f3724f6155')
-
-    @mock.patch.object(ResourceClient, 'get')
-    def test_get_with_uri_called_once(self, mock_get):
-        uri = '/rest/sas-interconnect-types/3518be0e-17c1-4189-8f81-83f3724f6155'
-        self._sas_interconnect_types.get(uri)
-
-        mock_get.assert_called_once_with(uri)